WIP refactor tokenization service
