+++ src/card/index.js
@@ -1,7 +1,5 @@
 const CardService = require('./card');
-const CardStatus = require('./card_status');
 
 module.exports = {
   CardService,
-  CardStatus,
 };
