+++ src/errors/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  JS_ERROR: 'JS_ERROR',
+  API_VALIDATION_ERROR: 'API_VALIDATION_ERROR',
+};
